--- a/app/assets/stylesheets/base.css
+++ b/app/assets/stylesheets/base.css
@@ -22,8 +22,12 @@ footer, header, hgroup, menu, nav, section {
 body {
   background-color: #e9573f;
   color: #000000;
-  font-family: "aktiv-grotesk-std", sans-serif;
+  font-family: "Helvetica Neue", Helvetica, Arial, sans-serif;
   font-size: 15px; }
+
+h1 {
+  font-family: Roboto, "Helvetica Neue", Helvetica, Arial, sans-serif;
+}
 
 section:not(:last-of-type) {
   border-bottom-width: 5px; }