--- a/app/assets/stylesheets/modules/button.css
+++ b/app/assets/stylesheets/modules/button.css
@@ -7,7 +7,6 @@
   border-radius: 20px;
   border: 1px solid transparent;
   font-weight: 600;
-  font-family: "aktiv-grotesk-std", sans-serif;
   font-size: 18px;
   cursor: pointer;
   outline: none;