--- a/app/assets/stylesheets/modules/form.css
+++ b/app/assets/stylesheets/modules/form.css
@@ -19,7 +19,6 @@
   width: 100%;
   font-weight: 300;
   font-size: 18px;
-  font-family: "aktiv-grotesk-std", sans-serif;
   border: 1px solid #f2f3f4;
   border-radius: 4px;
   box-shadow: inset 0 1px 1px -1px #c1c4ca;
@@ -60,7 +59,6 @@
   background-color: transparent;
   font-weight: 300;
   font-size: 18px;
-  font-family: "aktiv-grotesk-std", sans-serif;
   outline: none;
   z-index: 2;
   cursor: pointer;