--- a/app/assets/stylesheets/modules/org.css
+++ b/app/assets/stylesheets/modules/org.css
@@ -22,7 +22,6 @@
   border: 1px solid transparent;
   border-bottom: 1px solid #e9573f;
   font-weight: 500;
-  font-family: "aktiv-grotesk-std", sans-serif;
   font-size: 18px;
   background-image: linear-gradient(to right, rgba(233, 87, 63, 0.3), #e9573f);
   color: #e9573f;