As discussed in #1095, Typekit represents a huge amount of the total page weight of Rubygems.org. This PR aims to change that, while minimizing impact on the visual look and feel of the site.

This PR:

* Removes Typekit
* Swaps Aktiv Grotesk, the current site font, with a single "accent" h1 header font (Raleway) and uses system fonts for all remaining text.

## Philosophy

I think open source should be accessible to everyone - whether you have a snappy high bandwidth connection or not. Ruby is a global phenomenon, and I want to encourage even more people to try Ruby by making sure that our infrastructure performs well for people with average global connection speeds.

## Changes

Aktiv Grotesk was previously used for all site text. *Excluding* h1s, I have changed this to Helvetica Neue (Arial on non-Mac platforms). Here is a side-by-side of Aktiv Grotesk vs Helvetica, Aktiv on top:

![screen shot 2015-11-13 at 1 44 14 pm](https://cloud.githubusercontent.com/assets/845662/11155802/eb24e39e-8a15-11e5-9281-3c48d62e2f79.jpg)

The most visually distinctive typographic feature of the current design is the use of a light-weight font for h1s. I have retained that by swapping in Raleway, a similar humanist sans-serif. Here is a side-by-side of Aktiv Grotesk vs Raleway, Aktiv on top:

![screen shot 2015-11-13 at 1 55 22 pm](https://cloud.githubusercontent.com/assets/845662/11155828/0ed9941a-8a16-11e5-9add-46e8e206078c.jpg)

[Here's a comparison of a Gem page](http://imgur.com/a/FruOD).

## Benefits

Taking this approach shaves almost 400KB off the total page weight (almost 70%). That sounds nice, but what is really compelling is its impact on first-time page loads and paint times.

For this test, I used Google Chrome's network throttling feature. The site was running in production mode. I assumed a 625KB/s (5mbps) connection for the average global user (this number is pulled from Akamai), and a 1.5 MB/s (12mbps) connection for the average US user (also Akamai). 

### Current Live Site

| | Time to First Paint | Time to Paint Text (fonts loaded) | Time to Full Load |
| --- | --- | --- | --- |
| US  | 3.56s | 3.83s | 3.96s |
| Worldwide | 7.41s | 7.59s | 8.20s |

### This Pull Request

| | Time to First Paint | Time to Paint Text (fonts loaded) | Time to Full Load |
| --- | --- | --- | --- |
| US  | 854ms | 1.16s | 1.78s |
| Worldwide | 1.60s | 2.42s | 3.96s |

It's worth noting that, because there is no longer any render-blocking javascript before the CSS in the HEAD tag, the impact on paint times is far greater, percentage wise, than the impact on total `load` event times.

These numbers may be even better in production, as my local setup is slightly gimped versus the live site because I'm not serving the site assets gzipped.

**The average worldwide connection sees nearly a 75% improvement in time to painting all the text to the screen with this change.** Time to paint is more important than time to load in this case - load times are slightly behind the entire page actually being rendered.

## Open Questions

* Can we get a designer to review this change from a visual standpoint?
* Also a design question - should we consider adding Segoe UI or Calibri to the font stack for Windows users? Those fonts offer considerably more weights than Arial, but look slightly different. **EDIT**: No, let's leave that for another day.
* I used Google Fonts because Google will intelligently serve the best font format to each User Agent (WOFF2 to Chrome/FF, WOFF to IE9 >, I think TrueType for all else). We *could* inline a WOFF version into the CSS instead, which would allow us to eliminate Google entirely. There may be a small performance benefit to this. **EDIT**: Also leave for another day.