@RichiCoder1 2015-11-13
This is awesome, but could you not have simply moved the TypeKit javascript to a non-blocking position, and improved the font defaults so the pop-in wasn't as jarring?