@chills42 2015-11-13
Looks great! The only visual issue I see is with the gem version number in the title using raleway numbers.