@nateberkopec 2015-11-13
@RichiCoder1 Yes. However, we'd still be loading 14 different styles and weights of Aktiv Grotesk. So you wouldn't get the ~400kB page weight improvement you do here. Additionally, there aren't great fallback fonts for Windows or Android, and there is no appropriate fallback font for Linux.

@chills42 Good catch, Raleway appears prefer old-style baselines for numbers, rather than a modern "lining" style. [Whether or not this is a good thing is debatable](http://ux.stackexchange.com/questions/54740/what-is-the-appeal-of-old-style-numbers-and-why-do-modern-fonts-still-include-th). It's worth noting that the gem version number is intended to be in a light weight, so I think it's appropriate to keep it as Raleway. Whether or not we want lining numbers or old-style numbers, perhaps we can get a few more opinions on?