@gosukiwi 2015-11-14
I suggest using the default system font for each platform. 

`font-family: -apple-system, "Helvetica Neue", "Roboto", "Segoe UI", "Lucida Grande", sans-serif;`

Those are modern defaults for OS X, Android and Windows. [Medium](https://medium.com/@mwichary/system-shock-6b1dc6d6596f) uses that approach.