@bluebear94 2015-11-14
Yes for Segoe UI. No for Calibri (it looks ugly).