@arthurnn 2015-11-14
My intuition is to merge this right away, as this brings a big performance improvement and I like it 😸 .
I would like thoughts from @qrush first tho, before I merge, as he was more involved than me on the re-design of the website.
I am 👍 , and thanks a lot for the changes and tracing.