@nateberkopec 2015-11-14
Right, there's really only two open things:

* Should we change the main font stack to something other than what I've done? (Helvetica/Arial). I say *no*, as that was the original font stack, and changing the Windows fallback to something like Segoe or adding a more comprehensive fallback list is changing the original design too much. We can leave that issue for another day, it's orthogonal to the changes here really.
* Should we use lining numbers or old-style number for gem versions? Raleway uses "old-style" numbers by default: http://i.imgur.com/JooLgpm.png, note the baseline dropped on the 5. Changing this is a matter of setting [`font-variant-numeric`](https://developer.mozilla.org/en-US/docs/Web/CSS/font-variant-numeric).