@qrush 2015-11-20
I have a few questions:

1) Anyone from Dockyard have an opinion on this, given they led the design? @acacheung @bcardarella
2) What's the non-Latin character support for Raleway look like? Thinking Cyrillic, Chinese, Japanese, etc etc...

In general I am :+1: for this though. I would agree to not drop the baseline and set `font-variant-numeric` if we can change that.