@nateberkopec 2015-11-20
@qrush Raleway is available in *Latin only*. If this is a hard requirement, we can definitely investigate an alternative. Raleway just happens to be one that I think looks the closest to Aktiv Grotesk. Roboto, for an example, has Cyrillic, Greek and Latin Extended. Google Fonts does not publish any fonts with Japanese or Chinese character sets.

Worth noting that Aktiv Grotesk, as served by Typekit, does not support Japanese or Chinese characters. Confirmed in my browser by swapping my locale and checking the font.

Will update with some screenshots of lining numbers and an alternative font with extended character support.