@bcardarella 2015-11-20
I am 100% for moving to webfonts over Typekit. I am not happy that we (DockYard) went with Typekit to begin with and have since completely ditched it in favor of webfonts for all of our client projects. The design desire of the typography is outweighed here by the overwhelming evidence that Typekit is a performance nightmare.

:+1: