@nateberkopec 2015-11-23
I've decided to roll with Roboto over Raleway for language support reasons. I do not believe we have a locale file yet that can take advantage of this, but also I don't want someone to add a locale later with special characters (for example, an Eastern European language like Czech) and then be frustrated to see it not work with Raleway. Roboto looks close enough (though I am willing to be told be a designer otherwise!)

Here's Roboto on the homepage:

![screen shot 2015-11-23 at 2 40 42 pm](https://cloud.githubusercontent.com/assets/845662/11348213/1ec4bf50-91f3-11e5-8f84-71603aa533e9.png)

[Here's Roboto used to render the header of a gem's page](http://imgur.com/7AFibs4).

Advantages of swapping to Roboto: 

* Increases supportable languages to many Euro and non-European languages.
* We now have the same language support as we did when using Aktiv Grotesk.

Disadvantages of swapping to Roboto:

* We don't yet have a locale file that can fully utilize these character sets.
* This will increase font download size to about ~70kb in Firefox and in any other browsers that do not support [unicode-range](http://caniuse.com/#feat=font-unicode-range). Font download size is unaffected in Chrome and Opera, possibly Safari and IE, because those browsers can intelligently request which characters they need to render the page (they ignore the `subset` parameter in the Google Fonts URL in this PR). Total font download size is about ~10kb in these browsers. In any case, this is still a massive improvement over the current situation.
* It doesn't look *quite* as nice IMO as Raleway, but I'm willing to concede that it's probably bike shedding.

pinging @qrush