@simi 2015-11-23
@nateberkopec I think this can solve https://github.com/rubygems/rubygems.org/issues/732 also.

*ps.: I had an idea to translate rubygems.org into Czech. But everyone programming in Ruby can handle basic English in Czech republic, so I think it is not worth it.*

*ps2.: Czech is Central European language IMHO.*