@nateberkopec 2015-11-23
@simi Yes, hopefully we can close #783 and #785 with this! And sorry, I'm a silly American, you should feel lucky I even know the Czech Republic is in Europe! :stuck_out_tongue_closed_eyes: 

@dwradcliffe I sympathize, but through doing this PR, I've become more and more impressed with Google Fonts. UA-specific font delivery and other optimizations make it really difficult to compete against from a perf perspective! Of course, since [Roboto is Apache licensed](https://github.com/google/roboto/blob/master/LICENSE), we have a lot more leeway on messing with it ourselves though.