@qrush 2015-11-24
Big :+1:. Let's try it out.

I'm still not sure what to do about RDoc in readmes, geez that screenshot is rough.