@dwradcliffe 2015-11-24
I'm going to merge this and we can try it out. It looks like it will be a huge improvement and we can iterate on it if we want. Thanks @nateberkopec!