@nateberkopec 2015-11-24
:+1: I'll reinvestigate remaining open font issues once it's deployed