@simi 2015-11-24
Thanks! #732 is fixed for me.