@nateberkopec 2015-11-24
Looks like, as I expected, this is doing even better in production! 

### Current Production

| | Time to First Paint | Time to Paint Text (fonts loaded) | Time to Full Load |
| --- | --- | --- | --- |
| US (1.5 MB/s)  | 350ms | 510ms | 1.05s |
| Worldwide (625 KB/s) | 704ms | 1.27s | 2.17s |

That's a staggering **90% improvement in time to paint!** About an entire order of magnitude!